workedTimeをIntegerに統一
entityをdoubleに設定していたため、データ型変更によるＤＢ破壊を懸念して変更していなかったが、一度すべてのデータを消すタイミングがあったので変更。

+++ src/main/java/com/example/engineary/DTO/DiaryEntryResponse.java
@@ -20,7 +20,7 @@ public class DiaryEntryResponse {
     private String contents;
 
     /** 作業時間 */
-    private Double workedTime;
+    private Integer workedTime;
 
     /** 作業日時 デフォルトで当日 */
     private LocalDate workedDate;
