+++ src/main/java/com/example/engineary/model/DiaryEntry.java
@@ -34,7 +34,7 @@ public class DiaryEntry {
     private String contents;
 
     /** 作業時間 */
-    private Double workedTime;
+    private Integer workedTime;
 
     /** 作業日時 */
     private LocalDate workedDate;
